requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
langchain==0.2.7
langchain-community==0.2.6
chromadb==0.5.3
faiss-cpu==1.8.0.post1
pydantic==2.11.7
numpy==1.26.4
scikit-learn==1.7.1
pypdf==4.2.0
docx2txt==0.8
python-pptx==0.6.23
reportlab==4.1.0
requests==2.32.4
bcrypt==4.1.2
passlib==1.7.4
openai